+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All important changes to the decitala package will be documented here.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## ??
+### Added
+- Added a `DBParser.onset_ranges` method that returns a list of tuples holding the onset ranges of all extracted fragments
 
 ### Fixed
 - Issue #54 (unreferenced variable); Issue #51 (updated databases);